pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "darknet2any"
dynamic = ["version"]
description = "converts from darknet weights to tf, onnx, trt and other model formats"
requires-python = ">=3.8"

dependencies = [
  "ai_edge_litert",
  "numpy",
  "torch",
  "tensorboardX",
  "scikit_image",
  "matplotlib",
  "tqdm",
  "easydict",
  "Pillow",
  "opencv_python",
  "pycocotools",
  "simple_onnx_processing_tools",
  "onnx",
  "onnxruntime-gpu",
  "onnxsim-prebuilt",
  "onnx-graphsurgeon",
  "tensorflow",
  "onnx2tf",
  "tf-keras",
  "psutil",
  "h5py",
  "ml_dtypes",
]

[project.optional-dependencies]
tensorrt = [
  "nvidia-cublas-cu12",
  "nvidia-cuda-cupti-cu12",
  "nvidia-cuda-nvrtc-cu12",
  "nvidia-cuda-runtime-cu12",
  "nvidia-cudnn-cu12",
  "nvidia-cufft-cu12",
  "nvidia-curand-cu12",
  "nvidia-cusolver-cu12",
  "nvidia-cusparse-cu12",
  "nvidia-nccl-cu12",
  "nvidia-nvjitlink-cu12",
  "nvidia-nvtx-cu12",
  "tensorrt",
  "pycuda",
]

[project.scripts]
darknet2onnx = "darknet2any.darknet2onnx:main"
darknet2torch = "darknet2any.darknet2torch:main"
darknet2visual = "darknet2any.darknet2visual:main"
onnx2trt = "darknet2any.onnx2trt:main"
predict_darknet = "darknet2any.predict_darknet:main"
predict_onnx = "darknet2any.predict_onnx:main"
predict_tflite = "darknet2any.predict_tflite:main"
predict_trt = "darknet2any.predict_trt:main"
